Adicionado validações de parâmetros

+++ src/main/java/com/pds/biblioteca/service/firestore/AbstractFirestoreService.java
@@ -7,14 +7,14 @@ import com.google.cloud.firestore.DocumentSnapshot;
 import com.google.cloud.firestore.Firestore;
 import com.pds.biblioteca.entity.AbstractFirestoreEntity;
 import com.pds.biblioteca.exception.FirestoreExecuteException;
+import com.pds.biblioteca.util.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Optional;
+import java.util.*;
 import java.util.concurrent.ExecutionException;
 import java.util.stream.Collectors;
+
+import static java.util.Objects.isNull;
 
 public abstract class AbstractFirestoreService<T extends AbstractFirestoreEntity> {
 
@@ -27,6 +27,9 @@ public abstract class AbstractFirestoreService<T extends AbstractFirestoreEntity
     protected abstract String getCollectionName();
 
     protected DocumentReference addDocument(T obj) {
+        if (isNull(obj))
+            throw new FirestoreExecuteException("FIRESTORE_DATA_CREATION", "No data provided to save.");
+
         try {
             return this.firestore.collection(this.getCollectionName())
                     .add(this.toMap(obj))
@@ -51,10 +54,13 @@ public abstract class AbstractFirestoreService<T extends AbstractFirestoreEntity
         }
     }
 
-    protected void deleteDocument(String id) {
+    protected void deleteDocument(String documentId) {
+        if (StringUtils.isBlank(documentId))
+            throw new FirestoreExecuteException("FIRESTORE_DATA_DELETION", "Document id must not be empty.");
+
         try {
             this.firestore.collection(this.getCollectionName())
-                    .document(id)
+                    .document(documentId)
                     .delete()
                     .get();
         } catch (ExecutionException | InterruptedException e) {
@@ -63,9 +69,17 @@ public abstract class AbstractFirestoreService<T extends AbstractFirestoreEntity
     }
 
     protected void updateDocument(String documentId, T data) {
+        if (StringUtils.isBlank(documentId))
+            throw new FirestoreExecuteException("FIRESTORE_DATA_UPDATE", "Document id must not be empty.");
+
+        var map = removeNulls(toMap(data));
+
+        if (map.isEmpty())
+            throw new FirestoreExecuteException("FIRESTORE_DATA_UPDATE", "No data provided.");
+
         this.firestore.collection(this.getCollectionName())
                 .document(documentId)
-                .update(removeNulls(toMap(data)));
+                .update(map);
     }
 
     private Optional<T> fromDocument(DocumentSnapshot document, Class<T> cls) {
@@ -79,6 +93,9 @@ public abstract class AbstractFirestoreService<T extends AbstractFirestoreEntity
     }
 
     private Map<String, Object> toMap(T obj) {
+        if (obj == null)
+            return new HashMap<>(0);
+
         Map<String, Object> map = this.objectMapper.convertValue(obj, new TypeReference<>() {
         });
         map.remove("id");
